nova_ai/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
requests
aiosmtplib
diffusers
torch
transformers
accelerate
jinja2
python-dotenv
aiohttp
Pillow
